made core and neutron submodules
