+++ python/src/scipp/neutron/__init__.py
@@ -1,0 +1,7 @@
+# SPDX-License-Identifier: GPL-3.0-or-later
+# Copyright (c) 2019 Scipp contributors (https://github.com/scipp)
+# @file
+# @author Simon Heybrock
+
+from .._scipp.neutron import *
+from ..compat.mantid import load
